requirements.txt
# =========================
# Core Database Connectivity
# =========================
mysql-connector-python==8.3.0

# =========================
# Data Parsing & Utilities
# =========================
python-dateutil==2.9.0.post0

# =========================
# Logging & CLI Safety (Optional but Recommended)
# =========================
tqdm==4.66.2
# =========================
# For Tests
# =========================
pytest==8.1.1
# =========================
# Sentance Transformers
# =========================
sentence-transformers==2.6.1
# =========================
# Numpy
# =========================
numpy==1.26.4